Move 'build_comment' into the Phabricator reporter
The Phabricator reporter is the only thing using this function, and some
of the tasks contain Phabricator-flavoured markdown. So move this off
the base class for clarity.
