+++ bot/code_review_bot/report/phabricator.py
@@ -12,8 +12,47 @@ from libmozdata.phabricator import PhabricatorAPI
 from code_review_bot import Issue
 from code_review_bot import stats
 from code_review_bot.report.base import Reporter
+from code_review_bot.tasks.clang_tidy_external import ExternalTidyIssue
+from code_review_bot.tasks.coverage import CoverageIssue
+from code_review_tools import treeherder
 
 BUG_REPORT_URL = "https://bugzilla.mozilla.org/enter_bug.cgi?product=Firefox+Build+System&component=Source+Code+Analysis&short_desc=[Automated+review]+UPDATE&comment=**Phabricator+URL:**+https://phabricator.services.mozilla.com/...&format=__default__"
+
+COMMENT_FAILURE = """
+Code analysis found {defects_total} in the diff {diff_id}:
+"""
+
+COMMENT_RUN_ANALYZERS = """
+You can run this analysis locally with:
+{analyzers}
+"""
+
+COMMENT_COVERAGE = """
+In our previous code coverage analysis run, we found some files which had no coverage and are being modified in this patch:
+Should they have tests, or are they dead code?
+
+ - You can file a bug blocking [Bug 1415824](https://bugzilla.mozilla.org/show_bug.cgi?id=1415824) for untested files that should be **tested**.
+ - You can file a bug blocking [Bug 1415819](https://bugzilla.mozilla.org/show_bug.cgi?id=1415819) for untested files that should be **removed**.
+"""
+
+BUG_REPORT = """
+---
+If you see a problem in this automated review, [please report it here]({bug_report_url}).
+"""
+
+COMMENT_DIFF_DOWNLOAD = """
+For your convenience, [here is a patch]({url}) that fixes all the {analyzer} defects (use it in your repository with `hg import` or `git apply -p0`).
+"""
+
+COMMENT_TASK_FAILURE = """
+The analysis task [{name}]({url}) failed, but we could not detect any issue.
+Please check this task manually.
+"""
+
+FRONTEND_LINKS = """
+You can view these defects on [the code-review frontend]({frontend_url}) and on [Treeherder]({treeherder_url}).
+"""
+
 
 logger = structlog.get_logger(__name__)
 
@@ -122,3 +161,108 @@ class PhabricatorReporter(Reporter):
             ),
         )
         logger.info("Published phabricator summary")
+
+    def build_comment(
+        self,
+        revision,
+        issues,
+        bug_report_url,
+        frontend_url,
+        notices,
+        patches=[],
+        task_failures=[],
+    ):
+        """
+        Build a Markdown comment about published issues
+        """
+
+        def pluralize(word, nb):
+            assert isinstance(word, str)
+            assert isinstance(nb, int)
+            return "{} {}".format(nb, nb == 1 and word or word + "s")
+
+        # List all the issues classes
+        issue_classes = {issue.__class__ for issue in issues}
+
+        # Calc stats for issues, grouped by class
+        stats = self.calc_stats(issues)
+
+        # Build parts depending on issues
+        defects, analyzers = set(), set()
+        for stat in stats:
+            defect_nb = []
+            if stat["nb_build_errors"] > 0:
+                defect_nb.append(pluralize("build error", stat["nb_build_errors"]))
+            if stat["nb_defects"] > 0:
+                defect_nb.append(pluralize("defect", stat["nb_defects"]))
+
+            defects.add(
+                " - {nb} found by {analyzer}".format(
+                    analyzer=stat["analyzer"], nb=" and ".join(defect_nb)
+                )
+            )
+            _help = stat.get("help")
+            if _help is not None:
+                analyzers.add(f" - {_help}")
+
+        # Order both sets
+        defects = sorted(defects)
+        analyzers = sorted(analyzers)
+
+        # Build top comment
+        nb = len(issues)
+
+        if nb > 0:
+            comment = COMMENT_FAILURE.format(
+                defects_total=pluralize("defect", nb), diff_id=revision.diff_id
+            )
+        else:
+            comment = ""
+
+        # Add defects
+        if defects:
+            comment += "\n".join(defects) + "\n"
+
+        if analyzers:
+            comment += COMMENT_RUN_ANALYZERS.format(analyzers="\n".join(analyzers))
+
+        for tidy_external_issue in filter(
+            lambda i: isinstance(i, ExternalTidyIssue), issues
+        ):
+            comment += tidy_external_issue.as_markdown_for_phab()
+
+        for patch in patches:
+            comment += COMMENT_DIFF_DOWNLOAD.format(
+                analyzer=patch.analyzer.display_name, url=patch.url or patch.path
+            )
+
+        for task in task_failures:
+            treeherder_url = treeherder.get_job_url(
+                revision.repository_try_name,
+                revision.mercurial_revision,
+                task.id,
+                task.run_id,
+            )
+            comment += COMMENT_TASK_FAILURE.format(name=task.name, url=treeherder_url)
+
+        # Add coverage reporting details when a coverage issue is published
+        if CoverageIssue in issue_classes:
+            comment += COMMENT_COVERAGE
+
+        if notices:
+            # The '---' creates a horizontal rule in Phabricator's markdown
+            comment += "\n---\n".join(notices)
+
+        assert comment != "", "Empty comment"
+
+        comment += BUG_REPORT.format(bug_report_url=bug_report_url)
+
+        if defects:
+            treeherder_url = treeherder.get_job_url(
+                revision.repository_try_name, revision.mercurial_revision
+            )
+            comment += FRONTEND_LINKS.format(
+                frontend_url=frontend_url, treeherder_url=treeherder_url
+            )
+
+        return comment
